src/app/requirements.txt
dash-bootstrap-components>=1.5
plotly>=5.18
databricks-sdk>=0.20.0
